+++ src/Set.php
@@ -374,7 +374,7 @@ final class Set extends Collection
             // Validate the result of the callback.
             if (!is_bool($keep)) {
                 throw new UnexpectedValueException(
-                    'The filter callback must return a bool, got ' . Types::getBasicType($keep) . '.'
+                    'Cannot filter: callback returned ' . Types::getBasicType($keep) . ', expected bool.'
                 );
             }